+++ CHANGELOG.rst
@@ -10,7 +10,7 @@ Version 0.2
 Changes
 -------
 
-* Rebase from Girder 2.5 to Girder Unreleased
+* Rebase from Girder 2.5 to Girder 3.0 (Unreleased) | 2019-05-08
 
 Below is the changelog of the upstream Girder project's `|masterbranch| <https://github.com/girder/girder/tree/master>`_